requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
Pillow>=10.0.0
pdf2image>=1.16.0
